Reuse saveAnswer() and rename to upsertAnswer()
